+++ tools/who_what_benchmark/whowhatbench/inputs_preprocessors/minicpmo.py
@@ -4,7 +4,7 @@ from transformers import (
     PretrainedConfig,
     PreTrainedTokenizer,
 )
-from .preprocessors import VLMInputsPreprocessor
+from .vlm_inputs_preprocessor import VLMInputsPreprocessor
 from typing import TYPE_CHECKING, Optional, Union
 
 if TYPE_CHECKING:
@@ -12,9 +12,8 @@ if TYPE_CHECKING:
     from transformers.image_utils import VideoInput
 
 
-class Phi3MMInputsPreprocessor(VLMInputsPreprocessor):
+class MiniCPMOInputsPreprocessor(VLMInputsPreprocessor):
     def __init__(self, chat_mode: bool = False):
-        self.image_offset = 1
         super().__init__(chat_mode)
 
     def update_chat_history_with_answer(self, answer):
@@ -38,27 +37,21 @@ class Phi3MMInputsPreprocessor(VLMInputsPreprocessor):
             raise ValueError("Audio input is not supported")
 
         self.update_images(image)
-        if image is not None:
+        im_suffix = ""
+        if image:
             if not isinstance(image, list):
                 image = [image]
-            for i, _ in enumerate(image):
-                image_token = getattr(processor.tokenizer, "image_token", f"<|image_{i + self.image_offset}|>\n")
-                if image_token not in text:
-                    text = image_token + text
-            self.image_offset += len(image)
+            im_suffix = "(<image>./</image>)" * len(image) + "\n"
 
-        if getattr(processor.tokenizer, "chat_template", None) is None:
-            if self.chat_mode:
-                raise ValueError("Chat mode is not supported when there is no chat_template defined.")
+        new_message = {"role": "user", "content": im_suffix + text}
+        if self.chat_mode:
+            self.chat_history.append(new_message)
+            messages = self.chat_history
         else:
-            new_message = {"role": "user", "content": text}
-            if self.chat_mode:
-                self.chat_history.append(new_message)
-                chat_prompt = self.chat_history
-            else:
-                chat_prompt = [new_message]
+            messages = [new_message]
 
-            text = processor.tokenizer.apply_chat_template(chat_prompt, add_generation_prompt=True, tokenize=False)
+        prompt = processor.tokenizer.apply_chat_template(messages, tokenize=False, add_generation_prompt=True)
+        inputs = processor(prompt, self.images, return_tensors="pt")
+        inputs.pop("image_sizes", None)
 
-        inputs = processor(images=self.images, text=text, return_tensors="pt")
         return inputs
